# mern-delivery-modified/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Food Delivery API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  customer: 'customer',
  restaurant_owner: 'restaurant_owner',
  admin: 'admin',
} as const;

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role?: RegisterInputRole;
  phone?: string;
  address?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  customer: 'customer',
  restaurant_owner: 'restaurant_owner',
  admin: 'admin',
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  avatar?: string | null;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Restaurant {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  cuisine: string;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  coverImage?: string | null;
  address?: string;
  /** @nullable */
  phone?: string | null;
  /** Estimated delivery time in minutes */
  deliveryTime: number;
  deliveryFee: number;
  minOrder?: number;
  rating: number;
  reviewCount?: number;
  isOpen: boolean;
  /** @nullable */
  ownerId?: string | null;
  createdAt?: string;
}

export interface RestaurantInput {
  name: string;
  description?: string;
  cuisine: string;
  image?: string;
  coverImage?: string;
  address: string;
  phone?: string;
  deliveryTime: number;
  deliveryFee: number;
  minOrder?: number;
}

export interface RestaurantUpdate {
  name?: string;
  description?: string;
  cuisine?: string;
  image?: string;
  coverImage?: string;
  address?: string;
  phone?: string;
  deliveryTime?: number;
  deliveryFee?: number;
  minOrder?: number;
  isOpen?: boolean;
}

export interface RestaurantListResponse {
  restaurants: Restaurant[];
  total: number;
  page: number;
  limit: number;
}

export interface RestaurantStats {
  totalOrders: number;
  totalRevenue: number;
  avgRating: number;
  pendingOrders: number;
  todayOrders?: number;
  todayRevenue?: number;
}

export interface MenuCategory {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  restaurantId: string;
  sortOrder?: number;
}

export interface MenuCategoryInput {
  name: string;
  description?: string;
  sortOrder?: number;
}

export interface MenuCategoryUpdate {
  name?: string;
  description?: string;
  sortOrder?: number;
}

export interface MenuItem {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  image?: string | null;
  restaurantId: string;
  /** @nullable */
  categoryId?: string | null;
  isAvailable: boolean;
  isPopular?: boolean;
  isVegetarian?: boolean;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  prepTime?: number | null;
}

export interface MenuItemInput {
  name: string;
  description?: string;
  price: number;
  image?: string;
  categoryId?: string;
  isAvailable?: boolean;
  isPopular?: boolean;
  isVegetarian?: boolean;
  calories?: number;
  prepTime?: number;
}

export interface MenuItemUpdate {
  name?: string;
  description?: string;
  price?: number;
  image?: string;
  categoryId?: string;
  isAvailable?: boolean;
  isPopular?: boolean;
  isVegetarian?: boolean;
  calories?: number;
  prepTime?: number;
}

export interface CartItem {
  id: string;
  menuItemId: string;
  name: string;
  price: number;
  quantity: number;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Cart {
  /** @nullable */
  restaurantId?: string | null;
  /** @nullable */
  restaurantName?: string | null;
  items: CartItem[];
  subtotal: number;
  deliveryFee: number;
  total: number;
}

export interface CartItemInput {
  menuItemId: string;
  restaurantId: string;
  quantity: number;
  notes?: string;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface OrderItem {
  menuItemId: string;
  name: string;
  price: number;
  quantity: number;
  /** @nullable */
  image?: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  preparing: 'preparing',
  ready: 'ready',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: string;
  userId: string;
  restaurantId: string;
  restaurantName: string;
  items: OrderItem[];
  status: OrderStatus;
  subtotal: number;
  deliveryFee: number;
  total: number;
  deliveryAddress?: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  estimatedDelivery?: number | null;
  createdAt: string;
  updatedAt?: string;
}

export interface OrderInput {
  restaurantId: string;
  deliveryAddress: string;
  notes?: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  preparing: 'preparing',
  ready: 'ready',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
}

export interface Review {
  id: string;
  userId: string;
  userName: string;
  /** @nullable */
  userAvatar?: string | null;
  restaurantId: string;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  rating: number;
  comment?: string;
}

export interface PlatformStats {
  totalUsers: number;
  totalRestaurants: number;
  totalOrders: number;
  totalRevenue: number;
  activeRestaurants?: number;
  todayOrders?: number;
}

export interface PopularItem {
  menuItemId: string;
  name: string;
  restaurantName: string;
  orderCount: number;
  price: number;
  /** @nullable */
  image?: string | null;
}

export type ListRestaurantsParams = {
search?: string;
category?: string;
page?: number;
limit?: number;
};

export type ListMenuItemsParams = {
categoryId?: string;
};

export type ListOrdersParams = {
restaurantId?: string;
status?: string;
page?: number;
limit?: number;
};