# mern-delivery-modified/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Food Delivery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants/{restaurantId}/reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Post a review
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ── Platform Stats ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Review:
      type: object
      required: [id, userId, userName, restaurantId, rating, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        userAvatar:
          type: ["string", "null"]
        restaurantId:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
        comment:
          type: string

    # ── Stats ─────────────────────────────────────────────────────────────────
...